!F DEV - CS Buy/war rules
Immediately declaring war on a player after allying there city state (s) is not allowed.
This includes quest completions, election rigs and coups that replace the original ally.
This also includes allying the CS at the end of the turn and declaring war at the beginning of the next turn.
The original ally of the CS must always be given one full turn to react (After the last buy).
During this one turn, the original ally may buy back the CS and declare war.
The player attempting to take the CS from the original owner does not get this perk.